Changed drawing point-tolerance type to float refs #2582

+++ src/main/java/io/github/data4all/model/drawing/DrawingMotion.java
@@ -18,7 +18,7 @@ public class DrawingMotion {
 	/**
 	 * The default tolerance for a Point
 	 */
-	public static final int POINT_TOLERANCE = 5;
+	public static final float POINT_TOLERANCE = 5f;
 
 	private List<PointF> points = new ArrayList<PointF>();
